refactor: Repository 네이밍 변경(Room to Local) 및 Fake 구현체 도입을 통한 구조 개선
