+++ app/src/main/java/woowacourse/shopping/di/AppContainer.kt
@@ -7,9 +7,9 @@ import woowacourse.shopping.data.remote.HttpClientProvider
 import woowacourse.shopping.data.remote.MockWebServerProvider
 import woowacourse.shopping.data.remote.api.ProductService
 import woowacourse.shopping.data.remote.api.ProductServiceImpl
-import woowacourse.shopping.data.repository.cart.RoomCartRepository
+import woowacourse.shopping.data.repository.cart.LocalCartRepository
 import woowacourse.shopping.data.repository.product.RemoteProductRepository
-import woowacourse.shopping.data.repository.recent.RoomRecentProductRepository
+import woowacourse.shopping.data.repository.recent.LocalRecentProductRepository
 import woowacourse.shopping.domain.repository.CartRepository
 import woowacourse.shopping.domain.repository.ProductRepository
 import woowacourse.shopping.domain.repository.RecentProductRepository
@@ -31,7 +31,7 @@ class AppContainer(
             baseUrlProvider = { MockWebServerProvider.baseUrl },
         )
 
-    val cartRepository :CartRepository = RoomCartRepository(database.cartItemDao())
+    val cartRepository :CartRepository = LocalCartRepository(database.cartItemDao())
     val productRepository : ProductRepository = RemoteProductRepository(productService)
-    val recentProductRepository : RecentProductRepository = RoomRecentProductRepository(database.recentProductDao())
+    val recentProductRepository : RecentProductRepository = LocalRecentProductRepository(database.recentProductDao())
 }
